% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical.R
\name{critical_density}
\alias{critical_density}
\title{Calculate critical_density from free flow speed and jam density}
\usage{
critical_density(free_flow_speed = 130, jam_density = 150)
}
\arguments{
\item{free_flow_speed}{free flow speed in km/h (default 130 km/h)}

\item{jam_density}{jam density in vehicle per km (default = 150)}
}
\value{
Data frame of critical density containing: critical_density, speed_at_critical and max_flow
}
\description{
Calculate critical_density from free flow speed and jam density
}
\examples{
critical_density(free_flow_speed = 100)
critical_density(free_flow_speed = 130, jam_density = 150)
}
