% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{flow_from_density}
\alias{flow_from_density}
\title{Calculate traffic flow from density}
\usage{
flow_from_density(density, free_flow_speed = 130, jam_density = 150)
}
\arguments{
\item{density}{traffic density}

\item{free_flow_speed}{free flow speed in km/h (default 130 km/h)}

\item{jam_density}{jam density in vehicle per km}
}
\value{
Traffic flow in vehicles per hour
}
\description{
Calculate traffic flow from density
}
\examples{
flow_from_density(density = 10)
flow_from_density(density = c(10, 20, 30), free_flow_speed = 130)
}
