% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenshields.R
\name{greenshields}
\alias{greenshields}
\title{Calculation of the greenshields traffic flow model}
\usage{
greenshields(density, free_flow_speed = 130, jam_density = 150)
}
\arguments{
\item{density}{traffic density}

\item{free_flow_speed}{free flow speed in km/h (default 130 km/h)}

\item{jam_density}{jam density in vehicle per km}
}
\value{
Vector of speeds in km/h calculated from the Greenshields model
}
\description{
Calculation of the greenshields traffic flow model
}
\examples{
greenshields(density = 10)
greenshields(density = c(10, 20, 30), free_flow_speed = 130)
}
