% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_greenshields.R
\name{plot_greenshields}
\alias{plot_greenshields}
\title{Visualisation of the greenshields traffic flow model}
\usage{
plot_greenshields(free_flow_speed = 130, jam_density = 150)
}
\arguments{
\item{free_flow_speed}{free flow speed in km/h (default 130 km/h)}

\item{jam_density}{jam density in vehicle per km (default 150)}
}
\value{
A ggplot2 object showing speed-density relationship
}
\description{
Visualisation of the greenshields traffic flow model
}
\examples{
plot_greenshields(free_flow_speed = 100, jam_density = 100)
plot_greenshields(free_flow_speed = 130, jam_density = 150)
plot_greenshields(80, 100)
}
